+++ domain/src/main/java/com/netcracker_study_autumn_2020/domain/interactor/usecases/workspace/DeleteWorkspaceUseCase.java
@@ -8,5 +8,5 @@ public interface DeleteWorkspaceUseCase {
         void onError(Exception e);
     }
 
-    void execute(int workspaceId, DeleteWorkspaceUseCase.Callback callback);
+    void execute(long workspaceId, DeleteWorkspaceUseCase.Callback callback);
 }
